backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pinecone
langchain
langchain-community
langchain-huggingface
sentence-transformers
huggingface_hub